+++ src/backend/resolvers/mutations/person/create/index.js
@@ -1,0 +1,3 @@
+const createPerson = require('./create')
+
+module.exports = createPerson