+++ src/frontend/Orion/Configuration/AdminSettings/DevToProdPushManagement/DevToProdPushManagement.js
@@ -1,33 +1,35 @@
 import React from 'react'
 import { useQuery } from '@apollo/react-hooks'
 import styled from '@emotion/styled'
-import { lighten } from 'polished'
 
 import { GET_DEV_TO_PROD_PUSH_CONFIGS } from 'frontend/api/queries'
-import { Colors, Spacing } from 'frontend/utils/pulseStyles'
+import Color from 'frontend/utils/color'
+import Spacing from 'frontend/utils/spacing'
 
 import DataPushConfig from './DataPushConfig'
 import CreateDataPushConfigButton from './CreateDataPushConfigButton'
 import PushAllDataToProdButton from './PushAllDataToProdButton'
 
 const Wrapper = styled.div({
-  background: lighten(0.05, Colors.LIGHT_GRAY_1),
-  padding: Spacing.EXTRA_LARGE,
+  background: Color.GRAY_LIGHT,
+  padding: Spacing.S4,
   height: '100vh',
-  overflow: 'auto',
+  overflowY: 'auto',
   boxSizing: 'border-box',
-  flex: '1 0 auto',
+  flex: '1 1 auto',
 })
 
 const TopButtonsWrapper = styled.div({
   display: 'flex',
   justifyContent: 'flex-end',
+  padding: Spacing.S4,
   '& > :not(:last-child)': {
     marginRight: 12,
   },
 })
 
 const CardsContainer = styled.div({
+  padding: Spacing.S4,
   '& > :not(:last-child)': {
     marginBottom: 24,
   },