+++ src/frontend/api/queries.js
@@ -1,5 +1,29 @@
 import gql from 'graphql-tag'
 
+export const GET_EVENTS = gql`
+  query getEvents {
+    events {
+      _id
+      userId
+      username
+      action
+      entity
+      timestamp
+      boName
+      connectedEntities {
+        entity
+        boName
+      }
+      deltas {
+        field
+        before
+        after
+      }
+      metaType
+    }
+  }
+`
+
 export const GET_US_STATES = gql`
   query getUsStates {
     usStates {
@@ -80,6 +104,13 @@ export const GET_PEOPLE = gql`
       updatedOn
       firstName
       lastName
+      middleName
+      affiliation
+      affiliationPosition
+      primaryState
+      email
+      linkedIn
+      externalLink
       nationalProviderIdentifier
       physicianProfileId
     }